refactor(ui): break an import cycle
